+++ packages/components/psammead-most-read/src/List/index.jsx
@@ -8,7 +8,7 @@ import {
   WesternArabic,
 } from '@bbc/psammead-locales/numerals';
 import { scriptPropType } from '@bbc/gel-foundations/prop-types';
-import Grid from '../../../psammead-grid/esm';
+import Grid from '../../../psammead-grid';
 import { MostReadRank, MostReadLink } from '../Item';
 
 const MostReadItemProps = {
