+++ packages/components/psammead-most-read/src/testHelpers/stories.js
@@ -1,0 +1,17 @@
+import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
+
+export const arabicServiceDecorator = withServicesKnob({
+  defaultService: 'arabic',
+});
+
+export const bengaliServiceDecorator = withServicesKnob({
+  defaultService: 'bengali',
+});
+
+export const burmeseServiceDecorator = withServicesKnob({
+  defaultService: 'burmese',
+});
+
+export const newsServiceDecorator = withServicesKnob({
+  defaultService: 'news',
+});
